Extract db transaction utils from kraken-core
This exposes some transaction control features for Django which are
being adopted at Kraken, and are worth sharing more widely:

Atomicity:
- `transaction` (and `transaction_if_not_already`)
- `savepoint`
- `run_after_commit`

Constraints on function callers:
- `transaction_required`
- `durable`

Control flow:
- `in_transaction`
- `dbs_with_open_transactions`

Test utils:
- `part_of_a_transaction`
- escape-hatches for code still using `atomic` (only in pytest for now).

Using these, we have:

- explicit control over where transactions are opened.
- pressure to handle transactions mindfully.
- better after-commit callback simulation in tests.
- stricter after-commit callback registration.
- reduced incidental use of savepoints (and thus fewer queries).

This has required a little bit of a rewrite to remove off-topic features
(such as metrics), and opinionated choices (like always requiring the
"using" argument) which only apply in that codebase.

Perhaps we'll find a way to re-introduce those as features in this
codebase later, but this is a good first step. For now, those features
will continue to live in the parent codebase.

Co-authored-by: Lily Acorn <lily.acorn@kraken.tech>
Co-authored-by: Samuel Searles-Bryant <sam.searles-bryant@kraken.tech>