Control after-commit behaviour with a setting
By default, `run_after_commit` will raise an error if it is called when
it's not in a transaction. This could previously be controlled using a
test marker which monkey-patched a value in another module.

This new setting, `SUBATOMIC_AFTER_COMMIT_NEEDS_TRANSACTION`, allows
projects to disable this strict behaviour without resorting to
monkey-patching.

Co-authored-by: Samuel Searles-Bryant <sam.searles-bryant@kraken.tech>