Control after-commit calls in tests with a setting
By default, in (non transaction testcase) tests, we simulate the
execution of after-commit callbacks, to reproduce Django's normal
behaviour when outside of tests.

To allow projects to migrate their existing code over to this new
paradigm gradually, we allow this behaviour to be disabled. This was
previously controlled by monkey-patching a module scoped value in tests,
but is now exposed with a Django setting:
`SUBATOMIC_RUN_AFTER_COMMIT_CALLBACKS_IN_TESTS`.

Co-authored-by: Samuel Searles-Bryant <sam.searles-bryant@kraken.tech>