Remove implementation detail from public interface
This exception is raised when `savepoint` is used as a decorator, which
it doesn't support. The fact that it raises this exception is an
implementation detail we can expect to disappear in a future release.

Prepending an underscore to the name makes it clear that this is an
internal implementation detail, not a part of the public API. It also
means that it will not have automatically generated documentation.