Rework tests to remove uncovered lines
[Coverage] was rightly reporting that some lines in the tests were never
run.

This reworks the tests so that test branch coverage in tests is
complete. In most cases this means simply replacing `pass` statements
that were never expected to be reached with `...`. In one case, a `try`
statement is replaced with a boolean to track state. I feel that test is
improved by the change.

[Coverage]: https://coverage.readthedocs.io/