Rework how `savepoint` avoids use as a decorator
The old implementation subclassed a private object in `contextlib`, and
then partially broke some of its functionality.

This new version is more complicated, but less complex. It implements a
similar API to `contextlib.contextmanager`, but only produces a context
manager, and does not add support for use as a decorator.