Allow mypy to show color output
Before this change, the `TERM` environment variable was set to `dumb`,
which mypy interprets as being unable to display color. This means
that even the `FORCE_COLOR` environment variable does not force mypy to
output color.

This change sets the `TERM` environment variable to allow mypy to output
colors, making the mypy output easier to identify and interpret in CI.

See python/mypy#13817 for more information
about this problem.